Developed a web application for data visualization using data analysis and machine learning for balance scores of basketball players
Developed a system where users register and log in to access data on basketball coaches and players stored in a MySQL database. The home page features dropdown components allowing one-click access to basketball players' exercise data. A search bar is included for easy filtering of players, ensuring a user-friendly main page. The model is trained with three metrics: copx, copy, and balance score data of 5 basketball players, and is used to accurately predict data for 2 additional basketball players in the test set. The project includes specific data analysis graphs for the first 5 players and prediction graphs for the last 2 players, utilizing the random forest algorithm.